backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
black==22.1.0
boto3==1.21.13
botocore==1.24.13
click==8.0.3
Django==3.2
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
jmespath==0.10.0
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.4.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.0
typing-extensions==4.0.1
urllib3==1.26.8
whitenoise==6.0.0
